refactor: transfer verifyAction to userStore
