+++ src/screens/Login/LoginContainer.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 
-import { sampleLoginAction } from '~/store/modules/user/actions';
 import LoginViewer from './LoginViewer';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { NonLoginStackParams, NonLoginStackTypes } from '~/@types';
 import { stackNavigate } from '~/lib';
+import { sampleLoginAction } from '~/store/modules/auth/actions';
 
 type LoginProps = {
   navigation: StackNavigationProp<NonLoginStackParams, 'Login'>;
@@ -14,11 +14,7 @@ type LoginProps = {
 const LoginContainer = ({ navigation }: LoginProps) => {
   const dispatch = useDispatch();
   const handleLogin = () => {
-    dispatch(
-      sampleLoginAction({
-        isAuthenticated: true,
-      }),
-    );
+    dispatch(sampleLoginAction());
   };
   const handleNavigate = (to: NonLoginStackTypes) => {
     stackNavigate(navigation, to);
