+++ src/store/modules/auth/reducer.ts
@@ -1,15 +1,13 @@
-import { createReducer, action } from 'typesafe-actions';
+import { createReducer } from 'typesafe-actions';
 import { PostActions, AuthStateTypes } from './types';
 import {
   CREATE_ACCOUNT_SUCCESS,
   CREATE_ACCOUNT_FAILURE,
-  VERIFY_ID_SUCCESS,
-  VERIFY_ID_FAILURE,
-  VERIFY_NICKNAME_SUCCESS,
-  VERIFY_NICKNAME_FAILURE,
+  SAMPLE_LOGIN,
 } from './actions';
 
 const initialState: AuthStateTypes = {
+  isAuthenticated: false,
   user: {
     id: 0,
     identification: '',
@@ -19,10 +17,6 @@ const initialState: AuthStateTypes = {
   accessToken: {
     token: '',
     expire: 0,
-  },
-  duplicateCheck: {
-    isIdUsed: undefined,
-    isNicknameUsed: undefined,
   },
   refreshToken: '',
   error: '',
@@ -38,27 +32,9 @@ const authReducer = createReducer<AuthStateTypes, PostActions>(initialState, {
     ...state,
     ...action.payload,
   }),
-  [VERIFY_ID_SUCCESS]: (state, action) => {
-    const { status } = action.payload;
-    return {
-      ...state,
-      duplicateCheck: {
-        ...state.duplicateCheck,
-        isIdUsed: status !== 204 ? true : false,
-      },
-    };
-  },
-  [VERIFY_ID_FAILURE]: (state, action) => ({ ...state, ...action.payload }),
-  [VERIFY_NICKNAME_SUCCESS]: (state, action) => ({
+  [SAMPLE_LOGIN]: (state, action) => ({
     ...state,
-    duplicateCheck: {
-      ...state.duplicateCheck,
-      isNicknameUsed: action.payload.status !== 204 ? true : false,
-    },
-  }),
-  [VERIFY_NICKNAME_FAILURE]: (state, action) => ({
-    ...state,
-    ...action.payload,
+    isAuthenticated: true,
   }),
 });
