+++ src/store/modules/auth/types.ts
@@ -4,16 +4,12 @@ import { ActionType } from 'typesafe-actions';
 export type PostActions = ActionType<typeof actions>;
 
 export interface AuthStateTypes {
+  isAuthenticated: boolean;
   user: User;
   accessToken: AccessToken;
   refreshToken: string;
   error: string;
   status: number;
-  duplicateCheck: DuplicateCheck;
-}
-interface DuplicateCheck {
-  isIdUsed: undefined | boolean;
-  isNicknameUsed: undefined | boolean;
 }
 export interface AccessToken {
   token: string;
