+++ src/store/modules/user/thunks.ts
@@ -1,8 +1,13 @@
 import createAsyncThunk from '~/lib/createAsyncThunk';
 import callApi from '~/lib/callApi';
-import { postLoginRequestAsync } from './actions';
+import { verifyIdRequestAsync, verifyNicknameRequestAsync } from './actions';
 
-export const getPromotionListThunk = createAsyncThunk(
-  postLoginRequestAsync,
+export const postVerifyIdThunk = createAsyncThunk(
+  verifyIdRequestAsync,
+  callApi,
+);
+
+export const postVerifyNicknameThunk = createAsyncThunk(
+  verifyNicknameRequestAsync,
   callApi,
 );
